units: the writer emitted units it could not read back
A sweep of the whole unit space -- every base unit id crossed with every SI and
IEC prefix and every exponent, formatted and then parsed back -- turned up two
places where the library disagreed with itself, plus one in the bignum layer
that only a static analyser had been complaining about.

bvn_prefix_unit_valid range-checked the prefix ID in exactly one branch: SI
prefixes on bit and byte, where deciding "kilo and up" has to index the table to
compare magnitudes. Every other unit skipped it. That function is the single
gatekeeper the parser, the writer, the validator and the conversion helpers all
share, so an ID past the end of the table was "valid" in all of them at once.
si_prefix_str() has no symbol for such an ID and returns "", and
bvn_write_unit_component writes the '~' separator whenever a prefix is present
-- so a unit built with prefix.id.si = 99 on bu_meter formatted as "~m", a
separator with nothing in front of it. bvnr_write_float_unit accepted it and
reported success, and the document it produced,
.len=<float:64,~m>1.5e+0;, came back error_unit_illegal on the next read: a
silently unreadable document out of a writer that said it had succeeded.
bvn_unit_prefix_factor meanwhile scored the same unit as 1, dropping the scale
without a word, so the three disagreed three ways. The ID is now checked for
both prefix systems before anything looks at it, which is what makes them agree
again -- bvn_unit_valid says false, bvn_unit_to_string returns -1, and the
writer fails at the call instead of in somebody else's parser.

bvn_unit_equal called the two spellings of "no prefix" different units. An
absent prefix is either (prefix_si, si_none) -- what the parser always produces
-- or (prefix_iec, iec_none), which is what you reach for when you are building
an unprefixed bit or byte unit by hand. Nothing else distinguishes them: they
render to the same text and both score a factor of 1. But the component
comparison tested the prefix system before the ID, so a hand-built B compared
unequal to the B the parser hands back for the very text it prints. That is a
spurious unit mismatch for any caller that assembles its expected unit rather
than parsing one, and it contradicts this function's own documented contract
that equivalent notations compare equal. The absent-prefix case is now settled
before the system is allowed to matter; a real prefix still separates units, so
km and m are as unequal as they were.

bvn_int_mul, bvn_int_add and bvn_int_gcd size their scratch from the operand
widths -- na + nb, maxn + 1u -- and bvn_int_t is a public struct, so an operand
carrying a nonsensical nused made those sums wrap. A wrapped need then satisfied
bigint_ensure_cap's `need <= nlimbs` early return WITHOUT allocating, and the
multiply and add loops walked limbs that were never there. Nothing the library
builds looks like that, so this needed a hand-built or uninitialised struct
rather than one from bvn_int_alloc -- but these are BVN_API, and the same class
of unchecked field is what the unit bug above was. The three entry points screen
their operands, and bigint_ensure_cap no longer promises capacity for a buffer
that does not exist.

The unit fixes are pinned by tests that fail without them (16 failures with the
sources reverted and the tests kept). Also silences a -Wsign-conversion warning
in bvn_float_from_double, in both limb-width branches, where a `!= 0u` was being
converted against a `0u` ternary arm; the value was only ever 0 or 1 and nothing
about the rounding changes. A clean build is now warning-free with -fanalyzer.

+++ src/utils/bvn_float.c
@@ -1471,8 +1471,12 @@ bool bvn_float_from_double(bvn_float_t *f, double v)
 		unsigned dnum   = (unsigned)(-shift);          /* low bits dropped */
 		uint64_t kept   = full_man >> dnum;
 		uint64_t guard  = (full_man >> (dnum - 1u)) & 1u;
+		/* Both arms are uint64_t: a bare `!= 0u` yields int, which the usual
+		 * arithmetic conversions against `0u` would turn into an unsigned
+		 * conversion (-Wsign-conversion). The value is only ever 0 or 1. */
 		uint64_t sticky = (dnum >= 2u)
-			? ((full_man & (((uint64_t)1u << (dnum - 1u)) - 1u)) != 0u) : 0u;
+			? (uint64_t)((full_man & (((uint64_t)1u << (dnum - 1u)) - 1u)) != 0u)
+			: (uint64_t)0u;
 		if (guard && (sticky || (kept & 1u))) {
 			kept++;
 			if (kept >> (unsigned)prec) { kept >>= 1; f->_exp += 1; }
@@ -1492,8 +1496,12 @@ bool bvn_float_from_double(bvn_float_t *f, double v)
 		unsigned dnum   = (unsigned)(-shift);
 		uint64_t kept   = full_man >> dnum;
 		uint64_t guard  = (full_man >> (dnum - 1u)) & 1u;
+		/* Both arms are uint64_t: a bare `!= 0u` yields int, which the usual
+		 * arithmetic conversions against `0u` would turn into an unsigned
+		 * conversion (-Wsign-conversion). The value is only ever 0 or 1. */
 		uint64_t sticky = (dnum >= 2u)
-			? ((full_man & (((uint64_t)1u << (dnum - 1u)) - 1u)) != 0u) : 0u;
+			? (uint64_t)((full_man & (((uint64_t)1u << (dnum - 1u)) - 1u)) != 0u)
+			: (uint64_t)0u;
 		if (guard && (sticky || (kept & 1u))) {
 			kept++;
 			if (kept >> (unsigned)prec) { kept >>= 1; f->_exp += 1; }
